+++ src/test/resources/application.yml
@@ -14,3 +14,5 @@ spring:
       ddl-auto: create-drop
     defer-datasource-initialization: true
     open-in-view: false
+toss:
+  widget-secret-key: "test_gsk_docs_OaPz8L5KdmQXkzRz3y47BMw6"
